sago_assignment/requirements.txt
openai>=1.0.0
pypdf>=3.17.0
duckduckgo-search>=4.0.0
python-dotenv>=1.0.0
rich>=13.7.0
pytest>=8.0.0
reportlab>=4.0.0
pymongo>=4.0.0
